backend/requirements.txt
aiohttp==3.7.4
cchardet==2.1.7
gunicorn==20.1.0
httpie==3.0.2
PyJWT==1.7.0
httpx==0.22.0
aiohttp_swagger==1.0.4
motor==2.1.0
aio-pika==6.6.0
aiohttp_cors==0.6.0
aiobotocore==0.10.0
google-cloud
python-magic
libmagic
asyncio==3.4.2
Pillow
pytest
pytest-cov
pytest-mock
numpy
python-firebase
universal-analytics-python3==1.0.1
azure-storage-blob
azure-mgmt-storage
azure-servicebus
paramiko==2.7.2
scp==0.13.3